MaximallyLocalizedWannierFunctions

Documentation for MaximallyLocalizedWannierFunctions.

MaximallyLocalizedWannierFunctions.Hamiltonian

MaximallyLocalizedWannierFunctions.MLWF

MaximallyLocalizedWannierFunctions.MLWF

MaximallyLocalizedWannierFunctions.MLWF

MaximallyLocalizedWannierFunctions.δ

MaximallyLocalizedWannierFunctions.finitedifferences

MaximallyLocalizedWannierFunctions.gradient

MaximallyLocalizedWannierFunctions.r²ₙ

MaximallyLocalizedWannierFunctions.rₙ

MaximallyLocalizedWannierFunctions.Ω

MaximallyLocalizedWannierFunctions.Ω₁

MaximallyLocalizedWannierFunctions.Ω₂

QuantumLattices.update!

TightBindingApproximation.optimize!

MaximallyLocalizedWannierFunctions.Hamiltonian — Type

Hamiltonian{T<:Real, V<:AbstractVector{T}} <: Function

The Hamiltonian function based on Wannier interpolation.

MaximallyLocalizedWannierFunctions.MLWF — Type

MLWF(tba::TBA, brillouinzone::BrillouinZone, bands::AbstractVector{Int}, guess::Union{δ, Nothing}=nothing; kwargs...)
MLWF(hamiltonian::Function, lattice::Lattice, brillouinzone::BrillouinZone, bands::AbstractVector{Int}, guess::Union{δ, Nothing}=nothing; kwargs...)

Initialize a set of maximally localized Wannier functions.

MaximallyLocalizedWannierFunctions.MLWF — Type

MLWF{L<:Lattice, B<:BrillouinZone, T<:Real, V<:AbstractVector{T}} <: Function

Maximally localized Wannier functions on a Monkhorst-Pack mesh.

MaximallyLocalizedWannierFunctions.MLWF — Method

(mlwf::MLWF)(coordinate::AbstractVector) -> Matrix{Complex{dtype(mlwf)}}

Get the values of the wannier functions at each lattice point.

MaximallyLocalizedWannierFunctions.δ — Type

δ{T<:Real} <: Function

Delta function like Wannier functions.

MaximallyLocalizedWannierFunctions.finitedifferences — Method

finitedifferences(vectors::AbstractVector{<:AbstractVector}; nneighbor::Int=12, coordination::Int=12) -> Tuple{Vector, Vector{Vector}}

Get the weights and difference vectors of a Monkhorst-Pack discretization scheme whose minimum mesh translations are given by vectors.

See Sec. 3.2 in Reference for the details of the method.

MaximallyLocalizedWannierFunctions.gradient — Method

gradient(mlwf::MLWF) -> Vector{Matrix{Complex{dtype(mlwf)}}}

The gradient of the spread functional of the Wannier functions with respect to the infinitesimal gauge transformations at each momentum point.

MaximallyLocalizedWannierFunctions.r²ₙ — Method

r²ₙ(mlwf::MLWF) -> Vector{dtype(mlwf)}

⟨r²ₙ⟩, i.e., the expectation value of r²ₙ for each Wannier function.

MaximallyLocalizedWannierFunctions.rₙ — Method

rₙ(mlwf::MLWF) -> Vector{Vector{dtype(mlwf)}}

⟨rₙ⟩, i.e. the centers of the Wannier functions.

MaximallyLocalizedWannierFunctions.Ω — Method

Ω(mlwf::MLWF) -> dtype(mlwf)

The value of the spread functional of the Wannier functions.

MaximallyLocalizedWannierFunctions.Ω₁ — Method

Ω₁(mlwf::MLWF) -> dtype(mlwf)

The gauge-invariant part of the spread functional of the Wannier functions.

MaximallyLocalizedWannierFunctions.Ω₂ — Method

Ω₂(mlwf::MLWF) -> dtype(mlwf)

The gauge-dependent part of the spread functional of the Wannier functions.

QuantumLattices.update! — Method

update!(mlwf::MLWF, ::GradientDescent) -> MLWF

Update the maximally localized Wannier functions by the gradient descent method.

TightBindingApproximation.optimize! — Function

optimize!(mlwf::MLWF, method::FirstOrderOptimizer=GradientDescent(); maxiter::Int=20000, atol=atol, rtol=rtol, verbose::Union{Nothing, Int}=nothing) -> MLWF

Optimize a set of Wannier function to minimize the spread functional.